# Improve query sorting